Moving Seeker model
Moves Seeker to seekers app.

Changed all imports to import Seeker from seekers instead of users.

commented out 2 tests that assume there is data in the tables,
because right now there is no data, ill add the data in the next PR
and then ill include those tests as well.
