+++ users/models.py
@@ -3,7 +3,6 @@ from django.utils import timezone
 from django.contrib.auth.models import (AbstractBaseUser,
                                         PermissionsMixin,
                                         BaseUserManager)
-from roo_me.settings import AUTH_USER_MODEL
 
 
 class UserManager(BaseUserManager):
@@ -99,17 +98,3 @@ class User(AbstractBaseUser, PermissionsMixin):
             return self.apartment is not None
         except AttributeError:
             return False
-
-
-class Seeker(models.Model):
-    base_user = models.OneToOneField(AUTH_USER_MODEL, on_delete=models.CASCADE, primary_key=True)
-    city = models.ForeignKey('apartments.City', on_delete=models.RESTRICT, related_name='city_seekers')
-    start_date = models.DateField()
-    min_rent = models.IntegerField()
-    max_rent = models.IntegerField()
-    num_of_roomates = models.IntegerField()
-    num_of_rooms = models.IntegerField()
-    about = models.TextField(blank=True)
-
-    def __str__(self):
-        return f"Seeker:{self.base_user}"
